user intraction added

+++ src/app/modules/admin/admin.controller.ts
@@ -5,6 +5,8 @@ import { StatusCodes } from "http-status-codes";
 import { AdminService } from "./admin.service";
 import { getSingleFilePath } from "../../../shared/getFilePath";
 import { User } from "../user/user.model";
+import { Types } from "mongoose";
+import ApiError from "../../../errors/ApiError";
 
 
 const OverView = catchAsync(
@@ -43,7 +45,10 @@ const makeAPost = catchAsync(
 const deletePost = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
     const user = req.user;
-    const postID = req.query.id
+    const postID = req.query.id;
+    if (!postID) {
+      throw new ApiError(StatusCodes.BAD_REQUEST,"You must give the id to get the post")
+    }
     const result = await AdminService.deleteApost(user,postID as string);
 
     sendResponse(res, {
@@ -55,9 +60,118 @@ const deletePost = catchAsync(
   }
 );
 
+const subScriptions = catchAsync(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const user = req.user;
+    const id = req.query.id;
+    let result;
+    if (!id) {
+      result = await AdminService.getAllSubscriptions(user);
+    } else if ( typeof id === typeof Types.ObjectId ) {
+      result = await AdminService.ASubscription(user,id as string)
+    }
+
+    sendResponse(res, {
+      success: true,
+      statusCode: StatusCodes.OK,
+      message: 'Successfully get all subscription',
+      data: result,
+    });
+  }
+);
+
+// Hear will a create subscription system
+const createScriptions = catchAsync(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const user = req.user;
+    const result = await AdminService.getAllSubscriptions(user);
+
+    sendResponse(res, {
+      success: true,
+      statusCode: StatusCodes.OK,
+      message: 'Successfully create subscription',
+      data: result,
+    });
+  }
+);
+
+const editeScriptions = catchAsync(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const user = req.user;
+    const {...data} = req.body;
+    const result = await AdminService.updateSubscription(user,data);
+
+    sendResponse(res, {
+      success: true,
+      statusCode: StatusCodes.OK,
+      message: 'Successfully edited the subscription',
+      data: result,
+    });
+  }
+);
+
+const getUser = catchAsync(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const user = req.user;
+    const id = req.query.id;
+    let result;
+    if (!id) {
+      result = await AdminService.allUsers(user);
+    } else if ( id ) {
+      result = await AdminService.AUser(user,id as string)
+    }
+    sendResponse(res, {
+      success: true,
+      statusCode: StatusCodes.OK,
+      message: 'Successfully get user',
+      data: result,
+    });
+  }
+);
+
+const deletetUser = catchAsync(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const user = req.user;
+    const id = req.query.id;
+    if (!id) {
+      throw new ApiError(StatusCodes.BAD_REQUEST,"You must give the id of the user to delete the user")
+    }
+    const result = await AdminService.deleteUser(user,id as string)
+    sendResponse(res, {
+      success: true,
+      statusCode: StatusCodes.OK,
+      message: 'Successfully delete the user',
+      data: result,
+    });
+  }
+);
+
+const blockUser = catchAsync(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const user = req.user;
+    const id = req.query.id;
+    if (!id) {
+      throw new ApiError(StatusCodes.BAD_REQUEST,"You must give the id of the user to block the user")
+    }
+    const result = await AdminService.blockUser(user,id as string)
+    sendResponse(res, {
+      success: true,
+      statusCode: StatusCodes.OK,
+      message: 'Successfully block the user',
+      data: result,
+    });
+  }
+);
+
 
 export const AdminController = {
   OverView,
   makeAPost,
-  deletePost
+  deletePost,
+  subScriptions,
+  createScriptions,
+  editeScriptions,
+  getUser,
+  blockUser,
+  deletetUser
 }
